requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.6
django-environ==0.11.2
django-redis==5.4.0
django-rq==2.10.2
idna==3.7
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
redis==5.0.4
requests==2.32.3
rq==1.16.2
sqlparse==0.5.0
urllib3==2.2.1
